Duplicate components deleted. Routing changes. Folder layout changes and routes adjusted to match
