+++ src/app/creationsGallery/page.js
@@ -1,12 +1,15 @@
+import AllRecipes from "@/components/home/Recipes/allRecipes"
+import recipeFilters from "@/components/home/Recipes/recipeFilters"
+
 export default function creationsGallery(){
     return(
         <>
         <h1>Creations Gallery</h1>
         <div>
-            {/* Filters pane */}
+            <recipeFilters />
         </div>
         <div>
-            {/* Gallery tiles here */}
+            <AllRecipes />
         </div>
         </>
     )
