+++ src/app/page.js
@@ -1,8 +1,8 @@
 import { Flex, Text, Button } from "@radix-ui/themes";
 import "./home.css";
-import TopRecipes from "@/components/home/topRecipes";
-import MyCreations from "@/components/home/myCreations";
-import TopCreations from "@/components/home/topCreations";
+import TopRecipes from "@/components/home/Recipes/topRecipes";
+import MyCreations from "@/components/home/Creations/myCreations";
+import TopCreations from "@/components/home/Creations/topCreations";
 import Leaderboard from "@/components/home/leaderboard";
 import { db } from "@/utils/dbConnection";
