+++ src/components/home/Creations/newestCreations.jsx
@@ -1,13 +1,13 @@
 import { Card, Box, Separator, Heading, Text } from "@radix-ui/themes";
 
-export default function Leaderboard() {
+export default function NewestCreations() {
   return (
-    <div id="Leaderboard">
+    <div id="NewestCreations">
       <Box>
         <Card>
-          <Heading>Leaderboard</Heading>
+          <Heading>Newest Creations</Heading>
           <Separator my="3" size="4"></Separator>
-          <Text>Fill in for Leaderboard</Text>
+          <Text>Fill in for new Creations</Text>
         </Card>
       </Box>
     </div>
