+++ src/components/home/Creations/topCreations.jsx
@@ -1,13 +1,13 @@
 import { Card, Box, Separator, Heading, Text } from "@radix-ui/themes";
 
-export default function Leaderboard() {
+export default function TopCreations() {
   return (
-    <div id="Leaderboard">
+    <div id="TopCreations">
       <Box>
         <Card>
-          <Heading>Leaderboard</Heading>
+          <Heading>Top Creations</Heading>
           <Separator my="3" size="4"></Separator>
-          <Text>Fill in for Leaderboard</Text>
+          <Text>Fill in for top Creations</Text>
         </Card>
       </Box>
     </div>
