+++ src/components/home/Recipes/recipeFilters.jsx
@@ -1,13 +1,13 @@
 import { Card, Box, Separator, Heading, Text } from "@radix-ui/themes";
 
-export default function TopCreations() {
+export default function recipeFilters() {
   return (
-    <div id="TopRecipes">
+    <div id="Filters">
       <Box>
         <Card>
-          <Heading>Top Creations</Heading>
+          <Heading>Recipe Filters</Heading>
           <Separator my="3" size="4"></Separator>
-          <Text>Fill in for top Creations</Text>
+          {/* Filter pane goes here */}
         </Card>
       </Box>
     </div>
